expose the install._get_baseurl template in .teuthology.yaml
Instead of hardcoding the template used by _get_baseurl to build the URL
of the repository where packages are to be found, make it a

   baseurl_template

configuration parameter in the teuthology.yaml file.

_update_rpm_package_list_and_install did not call _get_baseurl and hard
coded another template identical to the one used by _get_baseurl. The
only difference was {distro_release} instead of {dist} but they are
expected to be identical for rpm based distributions.

http://tracker.ceph.com/issues/11830 Fixes: #11830

Signed-off-by: Loic Dachary <loic@dachary.org>