provision.py: implement an OpenStack backend
For machines of machine_type == openstack in paddles, when locking a
machine, an instance is created in the matching OpenStack cluster with:

   openstack server create redhat010

For instance, the machine redhat010 is created using the redhat OpenStack
cluster.

An OpenStack cluster is defined in the .teuthology.yaml file as follows:

openstack:
  user-data:
    ubuntu-14.04: teuthology/openstack-ubuntu-user-data.txt
    default: teuthology/openstack-default-user-data.txt
  default-size:
    disk-size: 10
    ram: 1024
    cpus: 1
  default-volumes:
    count: 1
    size: 1
  ssh-key: teuthology/test/integration/teuthology.pem
  clusters:
    the-re:
      openrc.sh: teuthology/test/integration/openrc.sh
      server-create: --availability-zone ovh
      volume-create: --availability-zone ovh --type ovh
      network: teuthology-test
      subnet: 10.50.50.0/24
      images:
        ubuntu-14.04: ubuntu-trusty-14.04
        centos-7.0: centos-7

When the machine is unlocked, it is destroyed with

   openstack server delete redhat010

The network is assumed to be exclusive to the OpenStack tenant and an
abitrary subnet can be chosen. The subnet can be connected to an existing
teuthology cluster via a VPN.

The python-openstackclient command line is used instead of the
corresponding API because it is well maintained and documented.

Integration tests require an OpenStack tenant.

Signed-off-by: Loic Dachary <loic@dachary.org>