provision.py: implement an OpenStack backend
For machines of machine_type == openstack in paddles, when locking a
machine, an instance is created in the matching OpenStack cluster with:

   openstack server create redhat

And renamed into redhat010 if assigned the IP x.x.x.10/24. It is then
locked in paddles which has been prepare with one slot for each
available IP in the range.

An OpenStack cluster is defined in the .teuthology.yaml file as follows:

openstack:
  user-data: teuthology/openstack-{os_type}-{os_version}-user-data.txt
  nameserver: 172.16.155.20
  default-size:
    disk-size: 10
    ram: 1024
    cpus: 1
  default-volumes:
    count: 1
    size: 1
  ssh-key: /home/ubuntu/.ssh/id_rsa
  clusters:
    redhat:
      openrc.sh: teuthology/test/integration/openrc.sh
      subnet: 172.16.155.0/24
      images:
        ubuntu-14.04: ubuntu-server-14.04-x86_64
        centos-7.0: Centos 7

When the machine is unlocked, it is destroyed with

   openstack server delete redhat010

The python-openstackclient command line is used instead of the
corresponding API because it is well maintained and documented.

Integration tests require an OpenStack tenant.

Signed-off-by: Loic Dachary <loic@dachary.org>